feat: add validation in pet register multistage so the user choose name and kind and species
