+++ src/components/Custom/PetRegister/PawIcon.tsx
@@ -15,39 +15,39 @@ export default function PawIcon({ step = 1, className }: PawIconProps) {
       xmlns="http://www.w3.org/2000/svg"
       className={className}
     >
-      <g clip-path="url(#clip0_4_73)">
+      <g clipPath="url(#clip0_4_73)">
         <path
           d="M33.6329 32.3696C35.557 31.2795 37.8805 30.7707 40.6134 30.9242C43.3365 31.0772 45.3786 31.8335 46.9884 33.0463C48.6108 34.2686 49.8819 36.0216 50.9464 38.3071V38.3061C52.4629 41.6621 55.061 44.4117 58.3244 46.1176V46.1186C61.1309 47.6745 63.246 50.2348 64.2443 53.2846C65.2423 56.3342 65.0504 59.6488 63.7072 62.5629C62.3638 65.4771 59.9682 67.7756 57.0011 68.9975C54.0405 70.2167 50.7293 70.2732 47.7287 69.1576C42.1974 66.7482 36.887 66.4517 31.254 69.1274L31.2423 69.1322L31.2316 69.1381C30.5925 69.4577 29.8337 69.6619 28.9533 69.8363C28.2249 69.9807 27.2086 70.1372 26.3693 70.3139C20.3677 69.9586 15.7083 66.3046 14.1564 60.5746L14.1525 60.561L14.0216 60.059C13.4219 57.5423 13.6378 54.8949 14.6456 52.5024C15.7197 49.9529 17.629 47.8444 20.0597 46.5238C23.8767 44.546 26.9665 41.4067 28.8839 37.559C30.1072 35.1737 31.7101 33.4589 33.6329 32.3696Z"
-          stroke-width="2.5"
+          strokeWidth="2.5"
           stroke="#ff7816"
         />
 
         <path
           d="M5.3877 22.8748C6.91583 22.2217 8.83016 22.2695 10.9297 23.3484C15.2045 25.5454 18.1718 30.4294 18.6162 36.1296C18.6422 37.8898 18.3646 39.4054 17.7383 40.6443C17.1181 41.871 16.123 42.8974 14.582 43.6394C13.0919 44.3569 11.7613 44.3888 10.4951 44.0251C9.18373 43.6484 7.88825 42.8305 6.55566 41.7517H6.55469C4.46019 40.0448 2.85716 37.5577 1.9834 34.8904C1.10707 32.2152 1.00001 29.4692 1.75 27.2478C2.51659 24.9775 3.85131 23.5313 5.3877 22.8748Z"
           stroke="#FF7816"
-          stroke-opacity= {step >= 2 ? "1" : "0.29"}
-          stroke-width="2.5"
+          strokeOpacity= {step >= 2 ? "1" : "0.29"}
+          strokeWidth="2.5"
         />
 
         <path
           d="M49.9726 2.48341C51.8844 1.29581 53.8881 1.01352 55.6142 1.46876C57.3324 1.92197 58.878 3.1317 59.8486 5.1299L59.8515 5.13575L59.8544 5.14259C60.3225 6.07862 60.6419 7.11478 60.8964 8.12501C61.0235 8.62941 61.1311 9.11243 61.2353 9.57423C61.3158 9.93067 61.3996 10.2835 61.4843 10.6065C60.8893 16.8034 59.3032 21.0942 55.0556 24.044C52.947 25.5082 50.9112 25.7928 49.1914 25.2949C47.453 24.7915 45.8562 23.435 44.7626 21.2774L44.7587 21.2705L44.622 20.9951C43.2491 18.1192 43.1334 14.4335 44.1025 10.9551C45.103 7.3641 47.2047 4.19093 49.9687 2.48536L49.9726 2.48341Z"
           stroke="#FF7816"
-          stroke-opacity= {step >= 4 ? "1" : "0.29"}
-          stroke-width="2.5"
+          strokeOpacity= {step >= 4 ? "1" : "0.29"}
+          strokeWidth="2.5"
         />
 
         <path
           d="M68.0115 23.408C70.18 22.2651 72.1347 22.2115 73.6746 22.8806C75.2234 23.5537 76.5634 25.0503 77.3035 27.4187L77.3045 27.4226C78.2028 30.2624 77.7036 33.7948 76.1561 36.9441C74.6085 40.0933 72.119 42.6426 69.3104 43.656C66.4494 44.6883 64.3413 44.1983 62.9295 42.8054C61.4416 41.3373 60.4534 38.6062 60.5106 34.6091C60.6826 32.2625 61.4581 29.9999 62.7635 28.0417C64.0753 26.074 65.8767 24.4812 67.99 23.4197L68.0115 23.408Z"
           stroke="#FF7816"
-          stroke-opacity= {step >= 5 ? "1" : "0.29"}
-          stroke-width="2.5"
+          strokeOpacity= {step >= 5 ? "1" : "0.29"}
+          strokeWidth="2.5"
         />
 
         <path
           d="M23.4947 1.38379C26.681 0.681725 31.1235 2.80494 33.442 7.38086C35.8281 12.0903 36.3833 16.3005 34.401 20.8057C32.5257 25.0675 27.6848 26.7137 23.5651 23.7227C19.1829 20.541 18.0105 16.4347 17.691 12.4326C17.6981 9.29965 18.3853 6.67253 19.4645 4.76367C20.5509 2.84214 21.9876 1.71606 23.4947 1.38379Z"
           stroke="#FF7816"
-          stroke-opacity= {step >= 3 ? "1" : "0.29"}
-          stroke-width="2.5"
+          strokeOpacity= {step >= 3 ? "1" : "0.29"}
+          strokeWidth="2.5"
         />
       </g>
       <defs>
